+++ android/src/main/kotlin/com/jarvan/fluwx/FluwxPlugin.kt
@@ -1,5 +1,7 @@
 package com.jarvan.fluwx
 
+import android.content.Intent
+import android.util.Log
 import androidx.annotation.NonNull
 import com.jarvan.fluwx.handlers.*
 import com.tencent.mm.opensdk.modelbiz.SubscribeMessage
@@ -16,16 +18,21 @@ import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
 
 /** FluwxPlugin */
-public class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
+class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware,PluginRegistry.NewIntentListener {
 
     companion object {
+
+        var callingChannel:MethodChannel? = null
+        // 主动获取的启动参数
+        var extMsg:String? = null
+
         @JvmStatic
         fun registerWith(registrar: PluginRegistry.Registrar) {
             val channel = MethodChannel(registrar.messenger(), "com.jarvanmo/fluwx")
             val authHandler = FluwxAuthHandler(channel)
-            FluwxResponseHandler.setMethodChannel(channel)
             WXAPiHandler.setContext(registrar.activity().applicationContext)
             channel.setMethodCallHandler(FluwxPlugin().apply {
+                this.fluwxChannel = channel
                 this.authHandler = authHandler
                 this.shareHandler = FluwxShareHandlerCompat(registrar).apply {
                     permissionHandler = PermissionHandler(registrar.activity())
@@ -38,15 +45,26 @@ public class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     private var authHandler: FluwxAuthHandler? = null
 
+    private var fluwxChannel: MethodChannel? = null
+
+    private fun handelIntent(intent:Intent?){
+        val action = intent?.action
+        val dataString = intent?.dataString
+        if (Intent.ACTION_VIEW == action) {
+            extMsg = dataString
+        }
+    }
+
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         val channel = MethodChannel(flutterPluginBinding.binaryMessenger, "com.jarvanmo/fluwx")
         channel.setMethodCallHandler(this)
-        FluwxResponseHandler.setMethodChannel(channel)
+        fluwxChannel = channel
         authHandler = FluwxAuthHandler(channel)
         shareHandler = FluwxShareHandlerEmbedding(flutterPluginBinding.flutterAssets, flutterPluginBinding.applicationContext)
     }
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+        FluwxPlugin.callingChannel = fluwxChannel
         when {
             call.method == "registerApp" -> WXAPiHandler.registerApp(call, result)
             call.method == "sendAuth" -> authHandler?.sendAuth(call, result)
@@ -60,6 +78,7 @@ public class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             call.method == "openWXApp" -> openWXApp(result)
             call.method.startsWith("share") -> shareHandler?.share(call, result)
             call.method == "isWeChatInstalled" -> WXAPiHandler.checkWeChatInstallation(result)
+            call.method == "getExtMsg" -> getExtMsg(result)
             else -> result.notImplemented()
         }
     }
@@ -75,16 +94,22 @@ public class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
         shareHandler?.permissionHandler = PermissionHandler(binding.activity)
+        handelIntent(binding.activity.intent)
     }
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
         WXAPiHandler.setContext(binding.activity.applicationContext)
+        handelIntent(binding.activity.intent)
         shareHandler?.permissionHandler = PermissionHandler(binding.activity)
     }
 
     override fun onDetachedFromActivityForConfigChanges() {
     }
 
+
+    private fun getExtMsg(result: MethodChannel.Result) {
+        result.success(extMsg)
+    }
 
     private fun pay(call: MethodCall, result: MethodChannel.Result) {
 
@@ -181,4 +206,9 @@ public class FluwxPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     }
 
     private fun openWXApp(result: MethodChannel.Result) = result.success(WXAPiHandler.wxApi?.openWXApp())
+
+    override fun onNewIntent(intent: Intent?): Boolean {
+        handelIntent(intent)
+        return false
+    }
 }