+++ lib/src/response/wechat_response.dart
@@ -40,19 +40,26 @@ Map<String, _WeChatResponseInvoker> _nameAndResponseMapper = {
       WeChatAuthGotQRCodeResponse.fromMap(argument),
   "onQRCodeScanned": (Map argument) =>
       WeChatQRCodeScannedResponse.fromMap(argument),
+  "onWXShowMessageFromWX": (Map argument) =>
+      WeChatShowMessageFromWXRequest.fromMap(argument),
 };
 
 class BaseWeChatResponse {
   final int errCode;
-  final String errStr;
+  final String? errStr;
 
   bool get isSuccessful => errCode == 0;
 
   BaseWeChatResponse._(this.errCode, this.errStr);
 
   /// create response from response pool
-  factory BaseWeChatResponse.create(String name, Map argument) =>
-      _nameAndResponseMapper[name](argument);
+  factory BaseWeChatResponse.create(String name, Map argument) {
+    var result = _nameAndResponseMapper[name];
+    if (result == null) {
+      throw ArgumentError("Can't found instance of $name");
+    }
+    return result(argument);
+  }
 }
 
 class WeChatShareResponse extends BaseWeChatResponse {
@@ -65,10 +72,10 @@ class WeChatShareResponse extends BaseWeChatResponse {
 
 class WeChatAuthResponse extends BaseWeChatResponse {
   final int type;
-  final String country;
-  final String lang;
-  final String code;
-  final String state;
+  final String? country;
+  final String? lang;
+  final String? code;
+  final String? state;
 
   WeChatAuthResponse.fromMap(Map map)
       : type = map["type"],
@@ -99,8 +106,8 @@ class WeChatAuthResponse extends BaseWeChatResponse {
 }
 
 class WeChatLaunchMiniProgramResponse extends BaseWeChatResponse {
-  final int type;
-  final String extMsg;
+  final int? type;
+  final String? extMsg;
 
   WeChatLaunchMiniProgramResponse.fromMap(Map map)
       : type = map["type"],
@@ -110,7 +117,7 @@ class WeChatLaunchMiniProgramResponse extends BaseWeChatResponse {
 
 class WeChatPaymentResponse extends BaseWeChatResponse {
   final int type;
-  final String extData;
+  final String? extData;
 
   WeChatPaymentResponse.fromMap(Map map)
       : type = map["type"],
@@ -119,10 +126,10 @@ class WeChatPaymentResponse extends BaseWeChatResponse {
 }
 
 class WeChatSubscribeMsgResponse extends BaseWeChatResponse {
-  final String openid;
-  final String templateId;
-  final String action;
-  final String reserved;
+  final String? openid;
+  final String? templateId;
+  final String? action;
+  final String? reserved;
   final int scene;
 
   WeChatSubscribeMsgResponse.fromMap(Map map)
@@ -135,9 +142,9 @@ class WeChatSubscribeMsgResponse extends BaseWeChatResponse {
 }
 
 class WeChatOpenBusinessWebviewResponse extends BaseWeChatResponse {
-  final int type;
+  final int? type;
   final int errCode;
-  final int businessType;
+  final int? businessType;
   final String resultInfo;
 
   WeChatOpenBusinessWebviewResponse.fromMap(Map map)
@@ -149,8 +156,8 @@ class WeChatOpenBusinessWebviewResponse extends BaseWeChatResponse {
 }
 
 class WeChatAuthByQRCodeFinishedResponse extends BaseWeChatResponse {
-  final String authCode;
-  final AuthByQRCodeErrorCode qrCodeErrorCode;
+  final String? authCode;
+  final AuthByQRCodeErrorCode? qrCodeErrorCode;
 
   WeChatAuthByQRCodeFinishedResponse.fromMap(Map map)
       : authCode = map["authCode"],
@@ -161,7 +168,7 @@ class WeChatAuthByQRCodeFinishedResponse extends BaseWeChatResponse {
 
 ///[qrCode] in memory.
 class WeChatAuthGotQRCodeResponse extends BaseWeChatResponse {
-  final Uint8List qrCode;
+  final Uint8List? qrCode;
 
   WeChatAuthGotQRCodeResponse.fromMap(Map map)
       : qrCode = map["qrCode"],
@@ -171,6 +178,15 @@ class WeChatAuthGotQRCodeResponse extends BaseWeChatResponse {
 class WeChatQRCodeScannedResponse extends BaseWeChatResponse {
   WeChatQRCodeScannedResponse.fromMap(Map map)
       : super._(map[_errCode], map[_errStr]);
+}
+
+// 获取微信打开App时携带的参数
+class WeChatShowMessageFromWXRequest extends BaseWeChatResponse {
+  final String? extMsg;
+
+  WeChatShowMessageFromWXRequest.fromMap(Map map)
+      : extMsg = map["extMsg"],
+        super._(0, '');
 }
 
 ///WechatAuth_Err_OK(0),